Factor out s:SetQuickfix() from s:ErrorsToQuickfix()
Replace s:GetCurrentQuickfixIdx() with different algorithm that counts the quickfix entries for a skipped buffer (by using the current index to go forward through the quickfix list until another buffer is encountered), avoiding the costly re-determining of the current index.
Clone quickfix entries of skipped buffers over to the fix command quickfix list.
s:DetermineSkippedEntries() already knows the skipped indices and just needs to return the entries for ArgsAndMore#Iteration#Quickfix() to accumulate. To keep the errors of non-skipped entries in the right position, these need to be interspersed; s:ConsumeErrors() does this during skipping, and once more after iteration for the remaining ones.